+++ terminal.c
@@ -2634,9 +2634,13 @@ term_xcursor_update(struct terminal *term)
 void
 term_set_window_title(struct terminal *term, const char *title)
 {
+    if (term->conf->locked_title && term->window_title_has_been_set)
+        return;
+
     free(term->window_title);
     term->window_title = xstrdup(title);
     render_refresh_title(term);
+    term->window_title_has_been_set = true;
 }
 
 void